refactor : seperate example and real code
